+++ crates/rattler_package_streaming/CHANGELOG.md
@@ -44,6 +44,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.13](https://github.com/conda/rattler/compare/rattler_package_streaming-v0.23.12...rattler_package_streaming-v0.23.13) - 2025-10-28
+
+### Other
+
+- updated the following local packages: rattler_conda_types, rattler_networking
+
+## [0.23.12](https://github.com/conda/rattler/compare/rattler_package_streaming-v0.23.11...rattler_package_streaming-v0.23.12) - 2025-10-25
+
+### Other
+
+- update Cargo.toml dependencies
+
 ## [0.23.11](https://github.com/conda/rattler/compare/rattler_package_streaming-v0.23.10...rattler_package_streaming-v0.23.11) - 2025-10-18
 
 ### Added